Set system time directly from hw clock in udev rule
With the new armmp kernel, RTC drivers are built as modules, and thus
the kernel doesn't set the system clock from the hardware clock when
the module is loaded, as explained in the RTC_HCTOSYS_DEVICE KConfig
option:

| The driver for this RTC device must be loaded before late_initcall
| functions run, so it must usually be statically linked.

In that case, util-linux should set the system clock from the RTC
itself. This is correctly done through /etc/init.d/hwclock.sh when udev
is not used. When using udev, the rule 85-hwclock.rules call
hwclock-set, which assumes that the system clock has already been set
earlier by the kernel and that the only remaining thing to do is the
correct it to the local timezone (--systz). While this was true with
the wheezy kernels, it's not longer true with the jessie one, and the
--systohc option has to be used instead.

Closes: #751238