Mangle installed files on sparc (sfdisk)
The sfdisk program is (intentionally) not built on sparc.

Instead of dropping sfdisk from generally installed files,
just mangle the installed files on sparc.

Note that this mangling is never cleaned up anywhere, so
noone should ever do an original upload on sparc but I
guess that it quite unlikely and we can expect that
a build on sparc is a build on a automated buildd.