Attempt to stop uuidd before usermod in postinst
If there is a running uuidd while executing the usermod command it will
error out. Attempt to stop it and hope that it doesn't get (socket)
activated before we get a chance to run the usermod command.
(Also hope that there are no other running processes under this
user for some funky reason.)
This should atleast catch most cases. Please note that uuidd is
not getting restarted.... Hopefully it'll get socket activated
if really needed. Also libuuid still should work even if uuidd is
not running. The user rename should happen so rarely that this should
not be an issue. Patches to improve this welcome.

LP: #1374648