Only set up system clock once
initramfs-tools [0.117] has started mounting and fscking both / and /usr in
the initramfs, and that means we need to set the system clock properly
as well. (#763777)

Currently /lib/udev/hwclock-set will happily adjust the system time
twice if the RTC uses local time and it is run in both the
initramfs and the main init system.

Closes: #763823