Avoid cyclic Breaks in initscripts vs util-linux
When util-linux breaks initscripts and initscripts breaks util-linux,
this happens:

E: This installation run will require temporarily removing the
   essential package initscripts:amd64 due to a
   Conflicts/Pre-Depends loop. This is often bad, but if you really
   want to do it, activate the APT::Force-LoopBreak option.
E: Internal Error, Could not early remove initscripts:amd64 (2)

Replaces should have a counterpart in either a versioned Breaks,
Conflicts or Depends. Bump the existing dependency and drop the Breaks
to avoid the cyclic Breaks to give apt the chance for a smooth upgrade.