Handle missing directory in obsolete conversion code
The migration should already have been done since wheezy->jessie
upgrade, but apparently this is an issue when host still runs wheezy and
schroot creates a >= jessie chroot which apparently copies the user
database from the host since it will run in a shared namespace.
You'll then end up with the old user (+group) but without the
old users home directory. That case is not handled correctly and
in normal conditions would never be triggered (ie. normally if the
user exists so would its home directory), which is probably why
noone noticed this until now.

This migration code is obsolete and should be dropped soon anyway
(just preserving the addgroup/adduser part of uuid-runtime.postinst)
but keep it around for the potential benefit of derivates that
has release schedules that doesn't match debians.

Closes: #804167