hwclock.sh: demote mountdevsubfs to Should-Start
As far as I can see hwclock does not require any sub-directories of /dev
to be mounted for proper operation, only /dev/rtc* et.al. needs
to be available (which in turn mountdevsubfs Requires udev for).
I find it very unlikely that it should fail and if /dev is not available
you likely have bigger problems than hwclock failing anyway.

The reason to demote it to Should-* instead of Required-* is to make
insserv not barf and error out if the required service is not available
at package install time. Consider a future system using for example
systemd (where the hwclock service is masked and not used as well)
then it might not be necessary to have the (sysvinit) 'initscripts'
package installed, but if it isn't the triggered insserv invokation
will fail if mountdevsubfs (from initscripts) is not available and
Required.

Since we're updating the hwclock init script (which is a configuration
file and thus user will be prompted if it's been locally modified)
prepare for the future case mentioned above to avoid needing
more config file merges than necessary.