Stop shipping the deprecated 'pg' utility
Upstream has deprecated this utility and set a deadline for its
removal. It's not built by default in upstreams build system
but we enabled it explicitly to postpone the removal until
early in a new Debian development cycle (instead of late
in the previous). It's time to bit the bullet.....