Move findmnt from mount to util-linux package
This is done in preparation for making the mount package non-essential.

(findmnt brings no new library dependencies to the util-linux package
and hopefully it's not problematic to build on non-Linux even though
it might not be very functional there because of libmount being
mostly stubs on non-Linux at the moment.)

Bump util-linux breaks/replaces on mount.
Add a versioned dependency on util-linux from mount package to
avoid broken upgrade combinations.