Split out cfdisk, fdisk and sfdisk to separate package
Splitting out the fdisk utilities into a separate package avoids
util-linux package to depend on libfdisk1.

Since these utilities where previously shipped in an essential package,
we need an upgrade plan. For now that plan is to have the util-linux
package depend on the new fdisk package to make fdisk pseudo-essential.
This should allow users of fdisk utilities to add the dependency and
once buster has been released (and hopefully all packages has added the
needed dependencies) we can demote the dependency to a recommends or
suggests and make fdisk uninstallable (by no longer being
pseudo-essential).

The new fdisk package gets demoted to priority important. Just for
safety add the Important: yes field to avoid anyone potentially
uninstalling the package by mistake for now. (We might want to
reconsider the Important: yes field later?)

This change is made to cater to the minimal chroot/container usecase
(in the future, as there's no practical difference right now).