Take over the rfkill package
Upstream merged rfkill tool into the util-linux suite. It was explicitly
mentioned by the (former) rfkill maintainer that he forgot to install
the package on a new Debian install and thus could not disable rfkill
(to be able to apt install rfkill), thus bumping the Priority field to
important (previously optional) to indicate it should be part of any
normal install. (But the ftp-masters will still need to update their
overrides file for the priority change to become effective.)

Potentially the rfkill tool should instead be shipped in the util-linux
binary package and the rfkill package be made a transitional package
instead.