Replace manual hwclock maintscript with debhelper generated
Move the linux-only-except-s390 special-casing from manually written
maintainerscripts (and overriding in debian/rules to not generate
any scripts for hwclock.sh) to instead doing the special-casing
in debian/rules and letting dh autogenerate maintainerscripts.
The autogenerated scripts are better than the manually written ones.

This has the minor side effect that the init script now doesn't
even get shipped on s390, but that could probably be considered
a bug before..... why have a init script on s390 that's disabled
by default because it won't work?! So lets consider that an
an additional bonus fix.
Left as an excercise for anyone who cares is to rm_conffile the
init script on s390(x) on upgrades.