Merge setpriv into util-linux package
This makes the command be part of the essential set, which is likely
needed as it's the recommended way to drop privilegies in a script
(on linux) when there's no need for a new PAM session to be opened up.

The command was previously in a separate package because it pulled
in new dependency libcap-ng0. The situation has possibly changed since
and other parts of the base system now pulls in libcap-ng0 anyway
(notably libaudit1 which is itself pulled in by libpam*, login,
etc....).
(The separate setpriv package was initially added in commit ad25bcb)

No transitional package should be needed as the setpriv package
doesn't seem to have any reverse dependencies in buster or sid
(plus it'll get pushed out by the Breaks/Replaces relation in
util-linux for those that have manually installed it).

(Note also that no special consideration was needed for stage1 builds
as the util-linux binary package is already completely excluded
from stage1.)