Make tests non-fatal on alpha and on arm-arm-* builders
The builders hosted at ARM Ltd have longstanding issues
that causes intermittent test failures.
See http://bugs.debian.org/872635
(Please note that this has nothing to do with the architecture in
question, rather in which data center it's located.)

The alpha architecture also has one test failing which has been
confirmed as a kernel bug by the alpha porters. Porters should
be working on getting it fixed, so let the build pass with this
known regression for now.