Add Build-Depends-Package to debian/*.symbols
This makes it possible for packages who want to generate stricter
dependencies than needed by the symbols they use to do so.  The version
in their build-dependency will be used if it's higher than any of the
versions listed for the used symbols when we specify this magic field.
Better have it before we actually need it, because by then it's to late
to add it.