Move all libraries to /usr/lib
Since we nowadays can rely on /usr being available during early boot,
there is no longer a need to install the libraries to /lib.
While at it, stop moving the libraries for the udeb build into a
non-multiarch location. This is not necessary as debian-installer
handles multi-arch just fine and this way we remove a pointless
difference between the udeb and non-udeb variant.
Drop --libdir and --libexecdir from CONFOPTS as debhelper sets those
properly for us.