Build with libcryptsetup-dev in dlopen mode to enable dm-verity
Re-enable dm-verity support. Use the new dlopen configure flag, so that
instead of linking against libcryptsetup.so, it is loaded dynamically
at runtime, and only if the verity feature options (-o verity...) are
used, to avoid the dependency on the libcryptsetup package being pulled
in unconditionally. Add it as a Suggests on cryptsetup-bin.

Gbp-Dch: full
Closes: #951048