Ship empty /var/lib/libuuid again
In 2014 we stopped creating this directory as part of the package. For
some more time, this was the home directory of the uuid runtime user,
but that got changed to /run/uuidd then. Ship this directory again,
as libuuid puts a file there, and since 2.38 the uuidd.service
definition also wants to bind-mount it.

Gbp-Dch: full
Closes: #1009686