d/control: Remove unused and seemingly invalid `Section: base`
The source stanza had a `Section: base` field, which was not used (all
package stanze seem to overwrite it) and referenced the non-existant
`base` section.

On the existance of the `base` section, Debian Policy Manual defers to
https://packages.debian.org/unstable/ as the canonical list of section
and that page does not have a `base` section. Though, even if it did
exist, the field is still redundant.

Signed-off-by: Niels Thykier <niels@thykier.net>