tests: mark fadvise/drop as known failing on buildds
Per discussion on #debian-buildd. The buildds have always used tmpfs as the
build location, and with schroot this is visible to the util-linux tests. They
in turn disable fadvise/drop, as it does not work on tmpfs.  With the unshare
backend, the actual backing filesystem is hidden, so the test gets run and
fails.

Gbp-Dch: full